% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{gbilinear_embed_spectral}
\alias{gbilinear_embed_spectral}
\title{Embed based on link function}
\usage{
gbilinear_embed_spectral(g, dim, f = log_odds, ...)
}
\description{
Estimates an embedding based on a bilinear link function.

The function
}
